SAM-2535 refactor authz checks

+++ samigo/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/listener/author/RemovePublishedAssessmentListener.java
@@ -46,6 +46,7 @@ import org.sakaiproject.tool.assessment.integration.helper.ifc.GradebookServiceH
 import org.sakaiproject.tool.assessment.services.assessment.PublishedAssessmentService;
 import org.sakaiproject.tool.assessment.ui.bean.author.AuthorBean;
 import org.sakaiproject.tool.assessment.ui.bean.author.PublishedAssessmentBean;
+import org.sakaiproject.tool.assessment.ui.bean.authz.AuthorizationBean;
 import org.sakaiproject.tool.assessment.ui.listener.util.ContextUtil;
 /**
  * <p>Title: Samigo</p>
@@ -58,10 +59,8 @@ public class RemovePublishedAssessmentListener
     implements ActionListener
 {
   private static Log log = LogFactory.getLog(RemovePublishedAssessmentListener.class);
-  private static final GradebookServiceHelper gbsHelper =
-		IntegrationContextFactory.getInstance().getGradebookServiceHelper();
-  private static final boolean integrated =
-		IntegrationContextFactory.getInstance().isIntegrated();
+  private static final GradebookServiceHelper gbsHelper = IntegrationContextFactory.getInstance().getGradebookServiceHelper();
+  private static final boolean integrated = IntegrationContextFactory.getInstance().isIntegrated();
   private CalendarServiceHelper calendarService = IntegrationContextFactory.getInstance().getCalendarServiceHelper();
 
   public RemovePublishedAssessmentListener()
@@ -77,8 +76,13 @@ public class RemovePublishedAssessmentListener
       log.debug("assessmentId = " + assessmentId); 	    
       PublishedAssessmentService assessmentService = new PublishedAssessmentService();
       //get assessment to see if it has a calendar event
-      PublishedAssessmentFacade assessment = assessmentService.getPublishedAssessment(
-				assessmentId.toString());
+      PublishedAssessmentFacade assessment = assessmentService.getPublishedAssessment(assessmentId.toString());
+
+      AuthorizationBean authzBean = (AuthorizationBean) ContextUtil.lookupBean("authorization"); 
+      if (!authzBean.isUserAllowedToDeleteAssessment(assessmentId, assessment.getCreatedBy(), true)) {
+        throw new IllegalArgumentException("User does not have permission to delete assessmentId " + assessmentId);
+      }
+
       assessmentService.removeAssessment(assessmentId, "remove");
       removeFromGradebook(assessmentId);
